(move) manifest test with manifest plugin
