+++ core/exports/effects.ts
@@ -1,6 +1,6 @@
-export { config, denoConfig } from "../src/effects/config.ts";
-export { env } from "../src/effects/env.ts";
-export { hot } from "../src/effects/hot-update.ts";
-export { importmap } from "../src/effects/importmap.ts";
-export { io } from "../src/effects/io.ts";
-export { manifest } from "../src/effects/manifest.ts";
+export { config, denoConfig } from "$effects/config.ts";
+export { env } from "$effects/env.ts";
+export { hot } from "$effects/hot-update.ts";
+export { importmap } from "$effects/importmap.ts";
+export { io } from "$effects/io.ts";
+export { manifest } from "$effects/manifest.ts";
