+++ core/exports/mod.ts
@@ -1,4 +1,4 @@
+export { importmapPath } from "$effects/importmap.ts";
+export { manifestPath } from "$effects/manifest.ts";
 export { globals } from "../src/constants.ts";
 export { startApp } from "../src/start.ts";
-export { manifestPath } from "../src/effects/manifest.ts";
-export { importmapPath } from "../src/effects/importmap.ts";
