+++ core/exports/plugins.ts
@@ -2,7 +2,10 @@ export { pluginConfig } from "../src/plugins/config.ts";
 export { pluginEnv } from "../src/plugins/env/env.ts";
 export { pluginImportmap } from "../src/plugins/importmap/importmap.ts";
 export { pluginIO } from "../src/plugins/io.ts";
-export { pluginManifest, updateManifest } from "../src/plugins/manifest.ts";
+export {
+  pluginManifest,
+  updateManifest,
+} from "../src/plugins/manifest/manifest.ts";
 export { pluginRender } from "../src/plugins/render/mod.ts";
 export { pluginStripTypes } from "../src/plugins/strip-types.ts";
 export { pluginBuild } from "../src/plugins/build.ts";
